requirements.txt
fastapi
uvicorn[standard]
python-dotenv
vllm
torch